separated dataset-specific munging tools, added preliminary comments to code
